renamed plot module to plots, renamed main small world module to bridges.py
